+++ src/future/mod.rs
@@ -1,0 +1,4 @@
+/// Provides the base client stubs used by the service macro.
+pub mod client;
+/// Provides the base server boilerplate used by service implementations.
+pub mod server;
